🔄 synced local 'skyvern/' with remote 'skyvern/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> This PR adds support for handling TOTP codes in actions, script generation, and task execution, including new parameters and logic for TOTP handling.
>
>   - **Behavior**:
>     - Add `totp_code_required` to `InputTextAction` in `actions.py`.
>     - Modify `_action_to_stmt()` in `generate_script.py` to handle TOTP parameters.
>     - Update `fill()`, `type()`, and `_input_text()` in `skyvern_page.py` to accept TOTP parameters.
>     - Introduce `poll_verification_code()` in `totp.py` for TOTP code retrieval.
>   - **Functions**:
>     - Update `parse_action()` in `parse_actions.py` to handle `totp_code_required`.
>     - Modify `handle_input_text_action()` in `handler.py` to process TOTP inputs.
>   - **Misc**:
>     - Remove redundant `poll_verification_code()` from `handler.py`.
>     - Add TOTP-related imports in `agent.py` and `parse_actions.py`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for b59cb83caeb07e56918ecefd0969b9f96455ab8a. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
